feat: implement card data getter for swipe-page

+++ server/API/Contracts/UserCardDto.cs
@@ -2,4 +2,4 @@
 
 namespace API.Contracts;
 
-public record UserCardDto(int Id, string Name, int Age, string Bio, double Distance, IEnumerable<string> Photos);
+public record UserCardDto(int Id, string Name, int Age, string Bio, int Distance, IEnumerable<string> Photos);
